requirements.txt
backtrader>=1.9.76.123
pandas>=1.5.3
numpy>=1.24.3
ta-lib>=0.4.24
quantstats>=0.0.57
akshare>=1.9.0
matplotlib>=3.7.1
futu-api